Refactor SendButton
