+++ DREAM Lab Frontend/src/pages/HomePage/components/SpeechBotCard/components/SendButton/SendButton.jsx
@@ -1,6 +1,8 @@
 import { useState, useEffect } from "react";
 
-import { API_URL } from "../Global/Database.js";
+import { API_URL } from "../../../../../../Global/Database.js";
+
+import "./SendButton.css";
 
 function SendButton({
     transcript,
